requirements.txt
streamlit>=1.32
openai>=1.54
python-dotenv>=1.0
requests>=2.31
